+++ ppcls/optimizer/__init__.py
@@ -21,7 +21,8 @@ import copy
 import paddle
 from typing import Dict, List
 
-from ..utils import logger, type_name
+from ppcls.engine.train.utils import type_name
+from ppcls.utils import logger
 
 from . import optimizer
 
@@ -44,10 +45,14 @@ def build_lr_scheduler(lr_config, epochs, step_each_epoch):
 
 
 # model_list is None in static graph
-def build_optimizer(config, max_iter, model_list, update_freq):
+def build_optimizer(engine):
+    if engine.mode != "train":
+        return None, None
+    config, max_iter, model_list = engine.config, engine.dataloader_dict[
+        "Train"].max_iter, [engine.model, engine.train_loss_func]
     optim_config = copy.deepcopy(config["Optimizer"])
     epochs = config["Global"]["epochs"]
-    update_freq = config["Global"].get("update_freq", 1)
+    update_freq = engine.update_freq
     step_each_epoch = max_iter // update_freq
     if isinstance(optim_config, dict):
         # convert {'name': xxx, **optim_cfg} to [{name: {scope: xxx, **optim_cfg}}]
