+++ ppcls/utils/__init__.py
@@ -26,8 +26,3 @@ from .metrics import multi_hot_encode
 from .metrics import precision_recall_fscore
 from .misc import AverageMeter
 from .save_load import init_model
-
-
-def type_name(object: object) -> str:
-    """get class name of an object"""
-    return object.__class__.__name__
